+++ internal/app/internal/orderer/orders/orderops.go
@@ -54,7 +54,7 @@ func (o *Orders) Balance(ctx context.Context, user string) (model.Balance, error
 	return balance, nil
 }
 
-func (o *Orders) Withdraw(ctx context.Context, user, order string, sum int) error {
+func (o *Orders) Withdraw(ctx context.Context, user, order string, sum float32) error {
 	order = compressOrderNumber(order)
 	if !checkOrderNumber(order) {
 		return ErrWrongOrderNumber
